test-requirements.txt
mysql-connector-python
pytest
pytest-mock
callee